requirements.txt
# Web server
fastapi==0.115.9
starlette==0.45.3
uvicorn==0.42.0
python-multipart==0.0.9
sse-starlette==2.3.6
websockets==14.2

# HTTP clients
httpx==0.28.1
requests==2.32.3
aiohttp==3.12.13
supabase==2.15.3
psycopg[binary]==3.2.3
psycopg-pool==3.2.3

# Config / env
python-dotenv==1.2.2
PyYAML==6.0.2
croniter==6.2.2
nest-asyncio==1.6.0
tenacity==9.1.2

# Auth
PyJWT==2.10.1
bcrypt==4.3.0
cryptography>=42.0.0

# LLM / AI
openai==1.102.0
anthropic==0.60.0
litellm==1.75.5.post1
openai-agents==0.2.10
tiktoken>=0.7.0

# MCP
mcp==1.26.0
fastmcp==2.12.2

# Tools / utilities
fire==0.7.1
click==8.1.8
rich==14.2.0
loguru==0.7.3
Jinja2==3.1.6
aiosqlite==0.21.0
validators>=0.22.0
defusedxml==0.7.1
PyNaCl==1.5.0

# Web research
firecrawl-py==3.2.0
exa-py==1.14.13
ddgs>=6.0.0
beautifulsoup4==4.13.4
lxml==6.0.0

# PDF / images
PyMuPDF==1.27.1
pillow==11.1.0
